Alteracao notfound no delete de Clieinte, Componente e Creatina

+++ creatinastore/src/main/java/br/com/creatinastore/Cliente/ClienteService.java
@@ -88,6 +88,10 @@ public class ClienteService {
 
     @Transactional
     public boolean delete(Long id) {
+        Cliente cliente = repository.findById(id);
+        if (cliente == null) {
+            throw new NotFoundException("Cliente com ID " + id + " não encontrado.");
+        }
         return repository.deleteById(id);
     }
 }
